Move all mem* functions to mem.c
