+++ fontforge/splinechar.c
@@ -31,6 +31,7 @@
 #include "fontforgevw.h"
 #include "fvfonts.h"
 #include "lookups.h"
+#include "mem.h"
 #include <math.h>
 #include <locale.h>
 # include <ustring.h>
