+++ fontforgeexe/mmdlg.c
@@ -30,6 +30,7 @@
 #include "fvfonts.h"
 #include "macenc.h"
 #include "mm.h"
+#include "mem.h"
 #include <ustring.h>
 #include <math.h>
 #include <gkeysym.h>
